requirements.txt
# TikTok Bot - Abhängigkeiten
requests>=2.31.0
python-dotenv>=1.0.0
moviepy>=1.0.3
gtts>=2.4.0
Pillow>=10.0.0
pytrends>=4.9.0
edge-tts>=6.0.0
